feat(event)!: Added onDetails event.
feat(event)!: Added onExhibitorCustomButtonClick event.

BREAKING CHANGE: To use the UIKit framework, the UIFplanView component was made. For SwiftUI it is better to use FplanView.

In this version of the package, the logic of loading the plan and caching has been changed. In the previous version of the package, a pre-prepared index.html file was opened and all resources were cached by intercepting all requests. In this version, the plan is opened via a direct link, and caching is carried out using WKWebView.